+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/InputAreaViewModel.kt
@@ -176,7 +176,7 @@ interface InputAreaViewModel {
     val useMarkdown: StateFlow<Boolean>
     val audio: AudioRecordingAreaViewModel
 
-    @Deprecated("use selectionMention with id instead") fun selectMention(userId: UserId) {}
+    @Deprecated("use selectMention with id instead") fun selectMention(userId: UserId) {}
 
     fun selectMention(id: String)
 
@@ -337,43 +337,42 @@ open class InputAreaViewModelImpl(
         override fun printHtml(html: CharSequence): CharSequence = html
     }
 
-    override val suggestedMentions: StateFlow<List<InputAreaViewModel.SuggestedMention>?> =
-        flow {
-                var lastEmptySearch: String? = null
-                fun calculateNextEmptySearch(currentEmptySearch: String?, listSize: Int, search: String): String? {
-                    return when {
-                        listSize != 0 -> null
-                        currentEmptySearch == null || currentEmptySearch.startsWith(search) -> search
-                        !search.startsWith(currentEmptySearch) -> null
-                        else -> currentEmptySearch
-                    }
-                }
-                emitAll(
-                    textField.map { textFieldValue ->
-                        val idLocalPartBeforeCursor = textFieldValue.mentionBeforeCursor()
-                        if (idLocalPartBeforeCursor != null) {
-                            lastEmptySearch?.let {
-                                if (idLocalPartBeforeCursor.startsWith(it)) {
-                                    return@map emptyList()
-                                }
-                            }
-
-                            _suggestedMentionsLoading.value = true
-                            val listOfMentions = buildList {
-                                addAll(listOfUsers(idLocalPartBeforeCursor))
-                                getRoomMentionIfAllowed(idLocalPartBeforeCursor)?.let { add(it) }
-                            }
-                            _suggestedMentionsLoading.value = false
-
-                            lastEmptySearch =
-                                calculateNextEmptySearch(lastEmptySearch, listOfMentions.size, idLocalPartBeforeCursor)
-
-                            listOfMentions
-                        } else null
-                    }
-                )
+    override val suggestedMentions: StateFlow<List<InputAreaViewModel.SuggestedMention>?> = flow {
+        var lastEmptySearch: String? = null
+        fun calculateNextEmptySearch(currentEmptySearch: String?, listSize: Int, search: String): String? {
+            return when {
+                listSize != 0 -> null
+                currentEmptySearch == null || currentEmptySearch.startsWith(search) -> search
+                !search.startsWith(currentEmptySearch) -> null
+                else -> currentEmptySearch
             }
-            .stateIn(coroutineScope, WhileSubscribed(), null)
+        }
+        emitAll(
+            textField.map { textFieldValue ->
+                val idLocalPartBeforeCursor = textFieldValue.mentionBeforeCursor()
+                if (idLocalPartBeforeCursor != null) {
+                    lastEmptySearch?.let {
+                        if (idLocalPartBeforeCursor.startsWith(it)) {
+                            return@map emptyList()
+                        }
+                    }
+
+                    _suggestedMentionsLoading.value = true
+                    val listOfMentions = buildList {
+                        addAll(listOfUsers(idLocalPartBeforeCursor))
+                        getRoomMentionIfAllowed(idLocalPartBeforeCursor)?.let { add(it) }
+                    }
+                    _suggestedMentionsLoading.value = false
+
+                    lastEmptySearch =
+                        calculateNextEmptySearch(lastEmptySearch, listOfMentions.size, idLocalPartBeforeCursor)
+
+                    listOfMentions
+                } else null
+            }
+        )
+    }
+        .stateIn(coroutineScope, WhileSubscribed(), null)
 
     override val listOfMentions: StateFlow<List<UserInfoElement>?> =
         suggestedMentions